Removed channel association rule
This should eliminate the false negative invalidations we're getting on litecoin for token trades. Also, if litecoin is spent and the tx is invalid that is very bad, that's like theft. So we need to not invalidate this easily. The tx expiration may be an exception to this principle, we will die of optionality otherwise. We'll set them high initially to play it safe and try to filter out optionality abuse.

We should also filter out the "channel not found" error, let's open this up for anyone to sell for ltc from a single sig address.