Removed multisig rule for Withdrawal
Being aware this may worsen and does not solve the issue of channel close not updating the commit registry properly, commented out the rule that the address one is withdrawing tokens from must be multisig. The check channel address function needs a look to make sure I'm not throwing anything valuable out. But this rule maybe should apply for commits if at all. Having a single-sig that one uses to do transfers is ok, but they need exception handling for trades (with only one address commiting, who is the counterparty? So maybe we can crash testnet trying that and add a rule to make it safe after verifying it's a problem.