No /0
Making it impossible to divide by 0. Not sure if 1 is 1 satoshi or 1 nominal unit of the denominator (vs. 1 satoshi of the denominator collateral) and if so it's still a liquidation monster in almost all cases. We're not going to have like Doge futures trading at 17 satoshis, hopefully. Technically someone could create a coin that's just for the lolz and scamming and have insane supply, decillions, and have it trade at 1 and 2 satoshis.

Anyways flaws aside this at least proofs against a crash if someone feeds a bad value. We can also put controls in the creation of oracle prints and contracts.