Commenting out the if brackets
This will only work for oracle contracts, not native contracts, likely to cause a consensus err if we use them. We'll need to have if logic for that.

For now, this makes it so the Oracle TWAP is the only source of price target for the market price order insertion.