Added refresh of position record when remainder = 0
Let's say I have 100 contracts and I sell 50. The code on lines 676-681 will execute. The record should reflect that the entry I am looking at, which has been reduced but not fully, is erased and replaced with an entry reflecting the new information. This plugs any value leak that might occur from counting that 50 contracts on another pass through of this logic.