Draft Realized PNL Logic
This one is double counting profits because it pays you the whole profit, entry to exit, and you've already been paid every settlement, so that needs to be unknotted. I need to somehow reference the mark price of the previous block to settle PNL for this blocks trade as the difference.